Fix reference to undefined command_name variable
In setup.py check_krb5_config checks for 'command_name' in kwargs and
sets cmd to that value or falls back to 'krb5-config'. If there is an
error when trying to run either of those commands it checks an undefined
command_name variable. It should be checking cmd instead.